get transaction and transaction detail feature

+++ controllers/productController.go
@@ -11,9 +11,14 @@ import (
 )
 
 func AddProductController(c echo.Context) error {
-  token := strings.Fields(c.Request().Header.Values("Authorization")[0])[1]
+	token := strings.Fields(c.Request().Header.Values("Authorization")[0])[1]
 	admin, err := middlewares.CheckTokenRole(token)
-  
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, map[string]string{
+			"message": err.Error(),
+		})
+	}
+
 	if admin {
 		var product models.Product
 
@@ -30,6 +35,7 @@ func AddProductController(c echo.Context) error {
 			"product": productResponse,
 		})
 	}
+
 	return c.JSON(http.StatusUnauthorized, map[string]string{
 		"message": "Unauthorized Action",
 	})
@@ -68,6 +74,9 @@ func GetProductDetailController(c echo.Context) error {
 func UpdateProductController(c echo.Context) error {
 	token := strings.Fields(c.Request().Header.Values("Authorization")[0])[1]
 	admin, err := middlewares.CheckTokenRole(token)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, map[string]string{
+			"message": err.Error(),
 
 	if admin {
 		var updatedProduct models.Product
@@ -104,8 +113,15 @@ func UpdateProductController(c echo.Context) error {
 }
 
 func DeleteProductController(c echo.Context) error {
-  token := strings.Fields(c.Request().Header.Values("Authorization")[0])[1]
-	admin, err := middlewares.CheckTokenRole(token)if admin {
+	token := strings.Fields(c.Request().Header.Values("Authorization")[0])[1]
+	admin, err := middlewares.CheckTokenRole(token)
+	if err != nil {
+		return c.JSON(http.StatusBadRequest, map[string]string{
+			"message": err.Error(),
+		})
+	}
+
+	if admin {
 		err := database.DeleteProduct(c.Param("id"))
 		if err != nil {
 			return c.JSON(http.StatusBadRequest, map[string]string{
